move RabitContext to rabit.py
