+++ tests/python-gpu/test_gpu_with_dask.py
@@ -405,7 +405,7 @@ class TestDistributedGPU:
             rabit_args = client.sync(dxgb._get_rabit_args, len(workers), None, client)
 
             def worker_fn(worker_addr: str, data_ref: Dict) -> None:
-                with dxgb.RabitContext(rabit_args):
+                with dxgb.DaskRabitContext(rabit_args):
                     local_dtrain = dxgb._dmatrix_from_list_of_parts(**data_ref, nthread=7)
                     fw_rows = local_dtrain.get_float_info("feature_weights").shape[0]
                     assert fw_rows == local_dtrain.num_col()
